+++ CHANGELOG.md
@@ -12,6 +12,7 @@ nav_order: 1
 
 ## [MAJOR.MINOR.PATCH] - YYYY-MM-DD
 
+- Fix: timeouts block values are now properly supported for resources implemented with the Plugin Framework
 - Promote `aiven_organization_billing_group` resource and data source, and `aiven_organization_billing_group_list` data
   source to general availability.
 - Deprecate `aiven_billing_group` resource and data source: use `aiven_organization_billing_group` instead.